tests: Unify single and multi record e2e tests
Combines the e2e and e2e-multi test suites into one unified test suite.
Tests for single and multi instance scenarios can still be executed
individually if required, but all tests should work in both cases.

`make test-e2e` - Runs all e2e tests
`make test-e2e-multi` - Runs only e2e tests for multiple record scenarios
