+++ test/e2e/multi_record_test.go
@@ -4,23 +4,26 @@ package e2e
 
 import (
 	"context"
+	"fmt"
 	"strings"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	. "github.com/onsi/gomega/gstruct"
+	"github.com/onsi/gomega/types"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	externaldnsendpoint "sigs.k8s.io/external-dns/endpoint"
 
 	"github.com/kuadrant/dns-operator/api/v1alpha1"
+	"github.com/kuadrant/dns-operator/internal/provider"
 	. "github.com/kuadrant/dns-operator/test/e2e/helpers"
 )
 
 // Test Cases covering multiple DNSRecords updating a set of records in a zone
-var _ = Describe("Multi Record Test", func() {
+var _ = Describe("Multi Record Test", Labels{"multi_record"}, func() {
 	// testID is a randomly generated identifier for the test
 	// it is used to name resources and/or namespaces so different
 	// tests can be run in parallel in the same cluster
@@ -30,17 +33,21 @@ var _ = Describe("Multi Record Test", func() {
 	// testHostname generated hostname for this test e.g. t-gw-mgc-12345.t-e2e-12345.e2e.hcpapps.net
 	var testHostname string
 
-	var dnsRecord1 *v1alpha1.DNSRecord
-	var dnsRecord2 *v1alpha1.DNSRecord
 	var geoCode1 string
 	var geoCode2 string
+
+	var testRecords []*testDNSRecord
+
+	recordsReadyMaxDuration := time.Minute
+	recordsRemovedMaxDuration := time.Minute
 
 	BeforeEach(func(ctx SpecContext) {
 		testID = "t-multi-" + GenerateName()
 		testDomainName = strings.Join([]string{testSuiteID, testZoneDomainName}, ".")
 		testHostname = strings.Join([]string{testID, testDomainName}, ".")
+		testRecords = []*testDNSRecord{}
 
-		if testDNSProvider == "gcp" {
+		if testDNSProvider == "google" {
 			geoCode1 = "us-east1"
 			geoCode2 = "europe-west1"
 		} else {
@@ -50,564 +57,478 @@ var _ = Describe("Multi Record Test", func() {
 	})
 
 	AfterEach(func(ctx SpecContext) {
-		if dnsRecord1 != nil {
-			err := k8sClient.Delete(ctx, dnsRecord1,
+		By("ensuring all dns records are deleted")
+		for _, tr := range testRecords {
+			err := tr.cluster.k8sClient.Delete(ctx, tr.record,
 				client.PropagationPolicy(metav1.DeletePropagationForeground))
 			Expect(client.IgnoreNotFound(err)).ToNot(HaveOccurred())
 		}
-		if dnsRecord2 != nil {
-			err := k8sClient.Delete(ctx, dnsRecord2,
-				client.PropagationPolicy(metav1.DeletePropagationForeground))
-			Expect(client.IgnoreNotFound(err)).ToNot(HaveOccurred())
-		}
+
+		By("checking all dns records are removed")
+		Eventually(func(g Gomega, ctx context.Context) {
+			for _, tr := range testRecords {
+				err := tr.cluster.k8sClient.Get(ctx, client.ObjectKeyFromObject(tr.record), tr.record)
+				g.Expect(err).To(MatchError(ContainSubstring("not found")))
+			}
+		}, time.Minute, 10*time.Second, ctx).Should(Succeed())
 	})
 
 	Context("simple", func() {
-		It("makes available a hostname that can be resolved", func(ctx SpecContext) {
-			By("creating two dns records")
-			testTargetIP1 := "127.0.0.1"
-			testTargetIP2 := "127.0.0.2"
-
-			dnsRecord1 = &v1alpha1.DNSRecord{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      testID + "-1",
-					Namespace: testNamespace,
-				},
-				Spec: v1alpha1.DNSRecordSpec{
-					RootHost: testHostname,
-					ManagedZoneRef: &v1alpha1.ManagedZoneReference{
-						Name: testManagedZoneName,
-					},
-					Endpoints: []*externaldnsendpoint.Endpoint{
-						{
-							DNSName: testHostname,
-							Targets: []string{
-								testTargetIP1,
-							},
-							RecordType: "A",
-							RecordTTL:  60,
+		It("creates and deletes distributed dns records", func(ctx SpecContext) {
+			By(fmt.Sprintf("creating %d simple dnsrecords accross %d clusters", len(testNamespaces)*len(testClusters), len(testClusters)))
+			for ci, tc := range testClusters {
+				for mi, mz := range tc.testManagedZones {
+					config := testConfig{
+						testTargetIP: fmt.Sprintf("127.0.%d.%d", ci+1, mi+1),
+					}
+					record := &v1alpha1.DNSRecord{
+						ObjectMeta: metav1.ObjectMeta{
+							Name:      testID,
+							Namespace: mz.Namespace,
 						},
-					},
-					HealthCheck: nil,
-				},
+						Spec: v1alpha1.DNSRecordSpec{
+							RootHost: testHostname,
+							ManagedZoneRef: &v1alpha1.ManagedZoneReference{
+								Name: mz.Name,
+							},
+							Endpoints: []*externaldnsendpoint.Endpoint{
+								{
+									DNSName: testHostname,
+									Targets: []string{
+										config.testTargetIP,
+									},
+									RecordType: "A",
+									RecordTTL:  60,
+								},
+							},
+							HealthCheck: nil,
+						},
+					}
+
+					By(fmt.Sprintf("creating dns record [name: `%s`, namespace: `%s`, mz: `%s`, endpoint: [dnsname: `%s`, target: `%s`]] on cluster [name: `%s`]", record.Name, record.Namespace, mz.Name, testHostname, config.testTargetIP, tc.name))
+					err := tc.k8sClient.Create(ctx, record)
+					Expect(err).ToNot(HaveOccurred())
+
+					testRecords = append(testRecords, &testDNSRecord{
+						cluster:     &testClusters[ci],
+						managedZone: mz,
+						record:      record,
+						config:      config,
+					})
+				}
 			}
 
-			dnsRecord2 = &v1alpha1.DNSRecord{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      testID + "-2",
-					Namespace: testNamespace,
-				},
-				Spec: v1alpha1.DNSRecordSpec{
-					RootHost: testHostname,
-					ManagedZoneRef: &v1alpha1.ManagedZoneReference{
-						Name: testManagedZoneName,
-					},
-					Endpoints: []*externaldnsendpoint.Endpoint{
-						{
-							DNSName: testHostname,
-							Targets: []string{
-								testTargetIP2,
-							},
-							RecordType: "A",
-							RecordTTL:  60,
-						},
-					},
-					HealthCheck: nil,
-				},
-			}
-
-			By("creating dnsrecord " + dnsRecord1.Name)
-			err := k8sClient.Create(ctx, dnsRecord1)
-			Expect(err).ToNot(HaveOccurred())
-
-			By("creating dnsrecord " + dnsRecord2.Name)
-			err = k8sClient.Create(ctx, dnsRecord2)
-			Expect(err).ToNot(HaveOccurred())
-
-			By("checking dns records become ready")
+			By(fmt.Sprintf("checking all dns records become ready within %s", recordsReadyMaxDuration))
 			Eventually(func(g Gomega, ctx context.Context) {
-				err := k8sClient.Get(ctx, client.ObjectKeyFromObject(dnsRecord1), dnsRecord1)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(dnsRecord1.Status.Conditions).To(
-					ContainElement(MatchFields(IgnoreExtras, Fields{
-						"Type":   Equal(string(v1alpha1.ConditionTypeReady)),
-						"Status": Equal(metav1.ConditionTrue),
-					})),
-				)
-				err = k8sClient.Get(ctx, client.ObjectKeyFromObject(dnsRecord2), dnsRecord2)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(dnsRecord2.Status.Conditions).To(
-					ContainElement(MatchFields(IgnoreExtras, Fields{
-						"Type":   Equal(string(v1alpha1.ConditionTypeReady)),
-						"Status": Equal(metav1.ConditionTrue),
-					})),
-				)
-			}, time.Minute, 10*time.Second, ctx).Should(Succeed())
+				for _, tr := range testRecords {
+					err := tr.cluster.k8sClient.Get(ctx, client.ObjectKeyFromObject(tr.record), tr.record)
+					g.Expect(err).NotTo(HaveOccurred())
+					g.Expect(tr.record.Status.Conditions).To(
+						ContainElement(MatchFields(IgnoreExtras, Fields{
+							"Type":   Equal(string(v1alpha1.ConditionTypeReady)),
+							"Status": Equal(metav1.ConditionTrue),
+						})),
+					)
+				}
+			}, recordsReadyMaxDuration, 5*time.Second, ctx).Should(Succeed())
 
-			By("checking dns records ownerID is set correctly")
-			Expect(dnsRecord1.Spec.OwnerID).To(BeEmpty())
-			Expect(dnsRecord2.Spec.OwnerID).To(BeEmpty())
-			Expect(dnsRecord1.Status.OwnerID).ToNot(BeEmpty())
-			Expect(dnsRecord2.Status.OwnerID).ToNot(BeEmpty())
-			Expect(dnsRecord1.Status.OwnerID).To(Equal(dnsRecord1.GetUIDHash()))
-			Expect(dnsRecord2.Status.OwnerID).To(Equal(dnsRecord2.GetUIDHash()))
-
-			testProvider, err := ProviderForManagedZone(ctx, testManagedZone, k8sClient)
+			By("checking provider zone records are created as expected")
+			testProvider, err := ProviderForManagedZone(ctx, testClusters[0].testManagedZones[0], testClusters[0].k8sClient)
 			Expect(err).NotTo(HaveOccurred())
 
-			By("ensuring zone records are created as expected")
-			Eventually(func(g Gomega, ctx context.Context) {
-				zoneEndpoints, err := EndpointsForHost(ctx, testProvider, testHostname)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(zoneEndpoints).To(HaveLen(2))
-				g.Expect(zoneEndpoints).To(ContainElements(
+			zoneEndpoints, err := EndpointsForHost(ctx, testProvider, testHostname)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(zoneEndpoints).To(HaveLen(2))
+
+			var allOwners = []string{}
+			var allTargetIps = []string{}
+			for i := range testRecords {
+				allOwners = append(allOwners, testRecords[i].record.Status.OwnerID)
+				allTargetIps = append(allTargetIps, testRecords[i].config.testTargetIP)
+			}
+
+			By("checking all target ips are present")
+			Expect(zoneEndpoints).To(ContainElements(
+				PointTo(MatchFields(IgnoreExtras, Fields{
+					"DNSName":       Equal(testHostname),
+					"Targets":       ConsistOf(allTargetIps),
+					"RecordType":    Equal("A"),
+					"SetIdentifier": Equal(""),
+					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
+				})),
+			))
+
+			By("checking all owner references are present")
+			for _, owner := range allOwners {
+				Expect(zoneEndpoints).To(ContainElements(
 					PointTo(MatchFields(IgnoreExtras, Fields{
-						"DNSName":       Equal(testHostname),
-						"Targets":       ConsistOf(testTargetIP1, testTargetIP2),
-						"RecordType":    Equal("A"),
-						"SetIdentifier": Equal(""),
-						"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-					})),
-					PointTo(MatchFields(IgnoreExtras, Fields{
-						"DNSName": Equal("kuadrant-a-" + testHostname),
-						"Targets": ContainElement(And(
-							ContainSubstring("heritage=external-dns,external-dns/owner="),
-							ContainSubstring(dnsRecord1.Status.OwnerID),
-							ContainSubstring(dnsRecord2.Status.OwnerID),
-						)),
+						"DNSName":       Equal("kuadrant-a-" + testHostname),
+						"Targets":       ContainElement(ContainSubstring(owner)),
 						"RecordType":    Equal("TXT"),
 						"SetIdentifier": Equal(""),
 						"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
 					})),
 				))
-			}, 10*time.Second, 1*time.Second, ctx).Should(Succeed())
+			}
 
-			By("ensuring the authoritative nameserver resolves the hostname")
+			By("checking the authoritative nameserver resolves the hostname")
 			// speed up things by using the authoritative nameserver
 			authoritativeResolver := ResolverForDomainName(testZoneDomainName)
 			Eventually(func(g Gomega, ctx context.Context) {
 				ips, err := authoritativeResolver.LookupHost(ctx, testHostname)
 				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(ips).To(ConsistOf(testTargetIP1, testTargetIP2))
+				g.Expect(ips).To(ContainElements(allTargetIps))
 			}, 300*time.Second, 10*time.Second, ctx).Should(Succeed())
 
-			By("deleting dnsrecord " + dnsRecord2.Name)
-			err = k8sClient.Delete(ctx, dnsRecord2,
+			//Test Deletion of one of the records
+			recordToDelete := testRecords[0]
+			lastRecord := len(testRecords) == 1
+			By(fmt.Sprintf("deleting dns record [name: `%s` namespace: `%s`]", recordToDelete.record.Name, recordToDelete.record.Namespace))
+			err = recordToDelete.cluster.k8sClient.Delete(ctx, recordToDelete.record,
 				client.PropagationPolicy(metav1.DeletePropagationForeground))
 			Expect(client.IgnoreNotFound(err)).ToNot(HaveOccurred())
 
-			By("checking dnsrecord " + dnsRecord2.Name + " is removed")
+			By(fmt.Sprintf("checking dns record [name: `%s` namespace: `%s`] is removed within %s", recordToDelete.record.Name, recordToDelete.record.Namespace, recordsRemovedMaxDuration))
 			Eventually(func(g Gomega, ctx context.Context) {
-				err := k8sClient.Get(ctx, client.ObjectKeyFromObject(dnsRecord2), dnsRecord2)
+				err := recordToDelete.cluster.k8sClient.Get(ctx, client.ObjectKeyFromObject(recordToDelete.record), recordToDelete.record)
 				g.Expect(err).To(HaveOccurred())
 				g.Expect(err).To(MatchError(ContainSubstring("not found")))
-			}, 25*time.Second, 1*time.Second, ctx).Should(Succeed())
+			}, recordsRemovedMaxDuration, 5*time.Second, ctx).Should(Succeed())
 
-			By("ensuring zone records are updated as expected")
+			By("checking provider zone records are updated as expected")
 			Eventually(func(g Gomega, ctx context.Context) {
 				zoneEndpoints, err := EndpointsForHost(ctx, testProvider, testHostname)
 				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(zoneEndpoints).To(HaveLen(2))
-				g.Expect(zoneEndpoints).To(ContainElements(
-					PointTo(MatchFields(IgnoreExtras, Fields{
-						"DNSName":       Equal(testHostname),
-						"Targets":       ConsistOf(testTargetIP1),
-						"RecordType":    Equal("A"),
-						"SetIdentifier": Equal(""),
-						"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-					})),
-					PointTo(MatchFields(IgnoreExtras, Fields{
-						"DNSName":       Equal("kuadrant-a-" + testHostname),
-						"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord1.Status.OwnerID + "\""),
-						"RecordType":    Equal("TXT"),
-						"SetIdentifier": Equal(""),
-						"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-					})),
-				))
-			}, 5*time.Second, 1*time.Second, ctx).Should(Succeed())
+				if lastRecord {
+					g.Expect(zoneEndpoints).To(HaveLen(0))
+				} else {
+					g.Expect(zoneEndpoints).To(HaveLen(2))
+					By(fmt.Sprintf("checking ip `%s` and owner `%s` are removed", recordToDelete.config.testTargetIP, recordToDelete.record.Status.OwnerID))
+					g.Expect(zoneEndpoints).To(ContainElements(
+						PointTo(MatchFields(IgnoreExtras, Fields{
+							"DNSName":       Equal(testHostname),
+							"Targets":       Not(ContainElement(recordToDelete.config.testTargetIP)),
+							"RecordType":    Equal("A"),
+							"SetIdentifier": Equal(""),
+							"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
+						})),
+						PointTo(MatchFields(IgnoreExtras, Fields{
+							"DNSName":       Equal("kuadrant-a-" + testHostname),
+							"Targets":       Not(ContainElement(ContainSubstring(recordToDelete.record.Status.OwnerID))),
+							"RecordType":    Equal("TXT"),
+							"SetIdentifier": Equal(""),
+							"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+						})),
+					))
+				}
+			}, 5*time.Second, time.Second, ctx).Should(Succeed())
 
-			By("ensuring the authoritative nameserver resolves the hostname")
+			By("deleting all remaining dns records")
+			for _, tr := range testRecords {
+				err := tr.cluster.k8sClient.Delete(ctx, tr.record,
+					client.PropagationPolicy(metav1.DeletePropagationForeground))
+				Expect(client.IgnoreNotFound(err)).ToNot(HaveOccurred())
+			}
+
+			By(fmt.Sprintf("checking all dns records are removed within %s", recordsRemovedMaxDuration))
 			Eventually(func(g Gomega, ctx context.Context) {
-				ips, err := authoritativeResolver.LookupHost(ctx, testHostname)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(ips).To(ConsistOf(testTargetIP1))
-			}, 300*time.Second, 10*time.Second, ctx).Should(Succeed())
+				for _, tr := range testRecords {
+					err := tr.cluster.k8sClient.Get(ctx, client.ObjectKeyFromObject(tr.record), tr.record)
+					g.Expect(err).To(MatchError(ContainSubstring("not found")))
+				}
+			}, recordsRemovedMaxDuration, 5*time.Second, ctx).Should(Succeed())
 
-			By("deleting dnsrecord " + dnsRecord1.Name)
-			err = k8sClient.Delete(ctx, dnsRecord1,
-				client.PropagationPolicy(metav1.DeletePropagationForeground))
-			Expect(client.IgnoreNotFound(err)).ToNot(HaveOccurred())
-
-			By("checking dnsrecord " + dnsRecord1.Name + " is removed")
-			Eventually(func(g Gomega, ctx context.Context) {
-				err := k8sClient.Get(ctx, client.ObjectKeyFromObject(dnsRecord1), dnsRecord1)
-				g.Expect(err).To(HaveOccurred())
-				g.Expect(err).To(MatchError(ContainSubstring("not found")))
-			}, 10*time.Second, 1*time.Second, ctx).Should(Succeed())
-
-			By("ensuring zone records are all removed as expected")
+			By("checking provider zone records are all removed")
 			Eventually(func(g Gomega, ctx context.Context) {
 				zoneEndpoints, err := EndpointsForHost(ctx, testProvider, testHostname)
 				g.Expect(err).NotTo(HaveOccurred())
 				g.Expect(zoneEndpoints).To(HaveLen(0))
-			}, 5*time.Second, 1*time.Second, ctx).Should(Succeed())
+			}, 5*time.Second, time.Second, ctx).Should(Succeed())
+
 		})
+
 	})
 
 	Context("loadbalanced", func() {
-		It("makes available a hostname that can be resolved", func(ctx SpecContext) {
+		It("creates and deletes distributed dns records", func(ctx SpecContext) {
 			if testDNSProvider == "azure" {
 				Skip("not yet supported for azure")
 			}
-			By("creating two dns records")
-			klbHostName := "klb." + testHostname
+			testGeoRecords := map[string][]testDNSRecord{}
 
-			testTargetIP1 := "127.0.0.1"
-			geo1KlbHostName := strings.ToLower(geoCode1) + "." + klbHostName
-			cluster1KlbHostName := "cluster1." + klbHostName
+			By(fmt.Sprintf("creating %d loadbalanced dnsrecords accross %d clusters", len(testNamespaces)*len(testClusters), len(testClusters)))
+			for ci, tc := range testClusters {
+				for mi, mz := range tc.testManagedZones {
 
-			testTargetIP2 := "127.0.0.2"
-			geo2KlbHostName := strings.ToLower(geoCode2) + "." + klbHostName
-			cluster2KlbHostName := "cluster2." + klbHostName
+					var geoCode string
+					if (ci+mi)%2 == 0 {
+						geoCode = geoCode1
+					} else {
+						geoCode = geoCode2
+					}
 
-			dnsRecord1 = &v1alpha1.DNSRecord{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      testID + "-1",
-					Namespace: testNamespace,
-				},
-				Spec: v1alpha1.DNSRecordSpec{
-					RootHost: testHostname,
-					ManagedZoneRef: &v1alpha1.ManagedZoneReference{
-						Name: testManagedZoneName,
-					},
-					Endpoints: []*externaldnsendpoint.Endpoint{
-						{
-							DNSName: cluster1KlbHostName,
-							Targets: []string{
-								testTargetIP1,
-							},
-							RecordType: "A",
-							RecordTTL:  60,
+					klbHostName := "klb." + testHostname
+					geoKlbHostName := strings.ToLower(geoCode) + "." + klbHostName
+					defaultGeoKlbHostName := strings.ToLower(geoCode1) + "." + klbHostName
+					clusterKlbHostName := fmt.Sprintf("cluster%d-%d.%s", ci+1, mi+1, klbHostName)
+
+					config := testConfig{
+						testTargetIP:       fmt.Sprintf("127.0.%d.%d", ci+1, mi+1),
+						testGeoCode:        geoCode,
+						testDefaultGeoCode: geoCode1,
+						hostnames: testHostnames{
+							klb:           klbHostName,
+							geoKlb:        geoKlbHostName,
+							defaultGeoKlb: defaultGeoKlbHostName,
+							clusterKlb:    clusterKlbHostName,
 						},
-						{
-							DNSName: testHostname,
-							Targets: []string{
-								klbHostName,
-							},
-							RecordType: "CNAME",
-							RecordTTL:  300,
+					}
+
+					record := &v1alpha1.DNSRecord{
+						ObjectMeta: metav1.ObjectMeta{
+							Name:      testID,
+							Namespace: mz.Namespace,
 						},
-						{
-							DNSName: geo1KlbHostName,
-							Targets: []string{
-								cluster1KlbHostName,
+						Spec: v1alpha1.DNSRecordSpec{
+							RootHost: testHostname,
+							ManagedZoneRef: &v1alpha1.ManagedZoneReference{
+								Name: mz.Name,
 							},
-							RecordType:    "CNAME",
-							RecordTTL:     60,
-							SetIdentifier: cluster1KlbHostName,
-							ProviderSpecific: externaldnsendpoint.ProviderSpecific{
+							Endpoints: []*externaldnsendpoint.Endpoint{
 								{
-									Name:  "weight",
-									Value: "200",
+									DNSName: clusterKlbHostName,
+									Targets: []string{
+										config.testTargetIP,
+									},
+									RecordType: "A",
+									RecordTTL:  60,
+								},
+								{
+									DNSName: testHostname,
+									Targets: []string{
+										klbHostName,
+									},
+									RecordType: "CNAME",
+									RecordTTL:  300,
+								},
+								{
+									DNSName: geoKlbHostName,
+									Targets: []string{
+										clusterKlbHostName,
+									},
+									RecordType:    "CNAME",
+									RecordTTL:     60,
+									SetIdentifier: clusterKlbHostName,
+									ProviderSpecific: externaldnsendpoint.ProviderSpecific{
+										{
+											Name:  "weight",
+											Value: "200",
+										},
+									},
+								},
+								{
+									DNSName: klbHostName,
+									Targets: []string{
+										geoKlbHostName,
+									},
+									RecordType:    "CNAME",
+									RecordTTL:     300,
+									SetIdentifier: config.testGeoCode,
+									ProviderSpecific: externaldnsendpoint.ProviderSpecific{
+										{
+											Name:  "geo-code",
+											Value: config.testGeoCode,
+										},
+									},
+								},
+								{
+									DNSName: klbHostName,
+									Targets: []string{
+										defaultGeoKlbHostName,
+									},
+									RecordType:    "CNAME",
+									RecordTTL:     300,
+									SetIdentifier: "default",
+									ProviderSpecific: externaldnsendpoint.ProviderSpecific{
+										{
+											Name:  "geo-code",
+											Value: "*",
+										},
+									},
 								},
 							},
+							HealthCheck: nil,
 						},
-						{
-							DNSName: klbHostName,
-							Targets: []string{
-								geo1KlbHostName,
-							},
-							RecordType:    "CNAME",
-							RecordTTL:     300,
-							SetIdentifier: geoCode1,
-							ProviderSpecific: externaldnsendpoint.ProviderSpecific{
-								{
-									Name:  "geo-code",
-									Value: geoCode1,
-								},
-							},
-						},
-						{
-							DNSName: klbHostName,
-							Targets: []string{
-								geo1KlbHostName,
-							},
-							RecordType:    "CNAME",
-							RecordTTL:     300,
-							SetIdentifier: "default",
-							ProviderSpecific: externaldnsendpoint.ProviderSpecific{
-								{
-									Name:  "geo-code",
-									Value: "*",
-								},
-							},
-						},
-					},
-					HealthCheck: nil,
-				},
+					}
+
+					By(fmt.Sprintf("creating dns record [name: `%s`, namespace: `%s`, managedZone: `%s`, endpoint: [dnsname: `%s`, target: `%s`, geoCode: `%s`]] on cluster [name: `%s`]", record.Name, record.Namespace, mz.Name, testHostname, config.testTargetIP, config.testGeoCode, tc.name))
+					err := tc.k8sClient.Create(ctx, record)
+					Expect(err).ToNot(HaveOccurred())
+					tr := &testDNSRecord{
+						cluster:     &testClusters[ci],
+						managedZone: mz,
+						record:      record,
+						config:      config,
+					}
+					testRecords = append(testRecords, tr)
+					testGeoRecords[config.testGeoCode] = append(testGeoRecords[config.testGeoCode], *tr)
+				}
 			}
 
-			dnsRecord2 = &v1alpha1.DNSRecord{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      testID + "-2",
-					Namespace: testNamespace,
-				},
-				Spec: v1alpha1.DNSRecordSpec{
-					RootHost: testHostname,
-					ManagedZoneRef: &v1alpha1.ManagedZoneReference{
-						Name: testManagedZoneName,
-					},
-					Endpoints: []*externaldnsendpoint.Endpoint{
-						{
-							DNSName: cluster2KlbHostName,
-							Targets: []string{
-								testTargetIP2,
-							},
-							RecordType: "A",
-							RecordTTL:  60,
-						},
-						{
-							DNSName: testHostname,
-							Targets: []string{
-								klbHostName,
-							},
-							RecordType: "CNAME",
-							RecordTTL:  300,
-						},
-						{
-							DNSName: geo2KlbHostName,
-							Targets: []string{
-								cluster2KlbHostName,
-							},
-							RecordType:    "CNAME",
-							RecordTTL:     60,
-							SetIdentifier: cluster2KlbHostName,
-							ProviderSpecific: externaldnsendpoint.ProviderSpecific{
-								{
-									Name:  "weight",
-									Value: "200",
-								},
-							},
-						},
-						{
-							DNSName: klbHostName,
-							Targets: []string{
-								geo2KlbHostName,
-							},
-							RecordType:    "CNAME",
-							RecordTTL:     300,
-							SetIdentifier: geoCode2,
-							ProviderSpecific: externaldnsendpoint.ProviderSpecific{
-								{
-									Name:  "geo-code",
-									Value: geoCode2,
-								},
-							},
-						},
-						//Note this dnsRecord has no default geo ....
-					},
-					HealthCheck: nil,
-				},
-			}
-
-			By("creating dnsrecord " + dnsRecord1.Name)
-			err := k8sClient.Create(ctx, dnsRecord1)
-			Expect(err).ToNot(HaveOccurred())
-
-			By("creating dnsrecord " + dnsRecord2.Name)
-			err = k8sClient.Create(ctx, dnsRecord2)
-			Expect(err).ToNot(HaveOccurred())
-
-			By("checking the dns records become ready")
+			By(fmt.Sprintf("checking all dns records become ready within %s", recordsReadyMaxDuration))
 			Eventually(func(g Gomega, ctx context.Context) {
-				err := k8sClient.Get(ctx, client.ObjectKeyFromObject(dnsRecord1), dnsRecord1)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(dnsRecord1.Status.Conditions).To(
-					ContainElement(MatchFields(IgnoreExtras, Fields{
-						"Type":   Equal(string(v1alpha1.ConditionTypeReady)),
-						"Status": Equal(metav1.ConditionTrue),
-					})),
-				)
-				err = k8sClient.Get(ctx, client.ObjectKeyFromObject(dnsRecord2), dnsRecord2)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(dnsRecord2.Status.Conditions).To(
-					ContainElement(MatchFields(IgnoreExtras, Fields{
-						"Type":   Equal(string(v1alpha1.ConditionTypeReady)),
-						"Status": Equal(metav1.ConditionTrue),
-					})),
-				)
-			}, time.Minute, 10*time.Second, ctx).Should(Succeed())
+				for _, tr := range testRecords {
+					err := tr.cluster.k8sClient.Get(ctx, client.ObjectKeyFromObject(tr.record), tr.record)
+					g.Expect(err).NotTo(HaveOccurred())
+					g.Expect(tr.record.Status.Conditions).To(
+						ContainElement(MatchFields(IgnoreExtras, Fields{
+							"Type":   Equal(string(v1alpha1.ConditionTypeReady)),
+							"Status": Equal(metav1.ConditionTrue),
+						})),
+					)
+				}
+			}, recordsReadyMaxDuration, 5*time.Second, ctx).Should(Succeed())
 
-			By("checking dns records ownerID is set correctly")
-			Expect(dnsRecord1.Spec.OwnerID).To(BeEmpty())
-			Expect(dnsRecord2.Spec.OwnerID).To(BeEmpty())
-			Expect(dnsRecord1.Status.OwnerID).ToNot(BeEmpty())
-			Expect(dnsRecord2.Status.OwnerID).ToNot(BeEmpty())
-			Expect(dnsRecord1.Status.OwnerID).To(Equal(dnsRecord1.GetUIDHash()))
-			Expect(dnsRecord2.Status.OwnerID).To(Equal(dnsRecord2.GetUIDHash()))
-
-			testProvider, err := ProviderForManagedZone(ctx, testManagedZone, k8sClient)
+			By("checking provider zone records are created as expected")
+			testProvider, err := ProviderForManagedZone(ctx, testClusters[0].testManagedZones[0], testClusters[0].k8sClient)
 			Expect(err).NotTo(HaveOccurred())
 
-			By("ensuring zone records are created as expected")
 			zoneEndpoints, err := EndpointsForHost(ctx, testProvider, testHostname)
 			Expect(err).NotTo(HaveOccurred())
-			if testDNSProvider == "gcp" {
-				Expect(zoneEndpoints).To(HaveLen(12))
-				//Main Records
-				By("checking endpoint " + testHostname)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(testHostname),
-					"Targets":       ConsistOf(klbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
-				By("checking endpoint " + klbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(klbHostName),
-					"Targets":       ConsistOf(geo1KlbHostName, geo2KlbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-					"ProviderSpecific": ContainElements(
-						externaldnsendpoint.ProviderSpecificProperty{Name: "routingpolicy", Value: "geo"},
-						externaldnsendpoint.ProviderSpecificProperty{Name: geo1KlbHostName, Value: geoCode1},
-						externaldnsendpoint.ProviderSpecificProperty{Name: geo2KlbHostName, Value: geoCode2},
-					),
-				}))))
-				By("checking endpoint " + geo1KlbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(geo1KlbHostName),
-					"Targets":       ConsistOf(cluster1KlbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-					"ProviderSpecific": ContainElements(
-						externaldnsendpoint.ProviderSpecificProperty{Name: "routingpolicy", Value: "weighted"},
-						externaldnsendpoint.ProviderSpecificProperty{Name: cluster1KlbHostName, Value: "200"},
-					),
-				}))))
-				By("checking endpoint " + geo2KlbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(geo2KlbHostName),
-					"Targets":       ConsistOf(cluster2KlbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-					"ProviderSpecific": ContainElements(
-						externaldnsendpoint.ProviderSpecificProperty{Name: "routingpolicy", Value: "weighted"},
-						externaldnsendpoint.ProviderSpecificProperty{Name: cluster2KlbHostName, Value: "200"},
-					),
-				}))))
-				By("checking endpoint " + cluster1KlbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(cluster1KlbHostName),
-					"Targets":       ConsistOf(testTargetIP1),
-					"RecordType":    Equal("A"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-				}))))
-				By("checking endpoint " + cluster2KlbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(cluster2KlbHostName),
-					"Targets":       ConsistOf(testTargetIP2),
-					"RecordType":    Equal("A"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-				}))))
-				//Txt Records
-				By("checking TXT owner endpoints")
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName": Equal("kuadrant-cname-" + testHostname),
-					"Targets": ContainElement(And(
+			var expectedEndpointsLen int
+			if testDNSProvider == "google" {
+				expectedEndpointsLen = (2 + len(testGeoRecords) + len(testRecords)) * 2
+				Expect(zoneEndpoints).To(HaveLen(expectedEndpointsLen))
+			} else if testDNSProvider == "aws" {
+				expectedEndpointsLen = (2 + len(testGeoRecords) + (len(testRecords) * 2)) * 2
+				Expect(zoneEndpoints).To(HaveLen(expectedEndpointsLen))
+			}
+
+			var totalEndpointsChecked = 0
+			var allOwners = []string{}
+			var allOwnerMatcher = []types.GomegaMatcher{
+				ContainSubstring("heritage=external-dns,external-dns/owner="),
+			}
+			var geoOwners = map[string][]string{}
+			var geoKlbHostname = map[string]string{}
+			var geoOwnerMatcher = map[string][]types.GomegaMatcher{}
+			for i := range testRecords {
+				ownerID := testRecords[i].record.Status.OwnerID
+				allOwners = append(allOwners, ownerID)
+				allOwnerMatcher = append(allOwnerMatcher, ContainSubstring(ownerID))
+
+				geoCode := testRecords[i].config.testGeoCode
+				geoOwners[geoCode] = append(geoOwners[geoCode], ownerID)
+				geoKlbHostname[geoCode] = testRecords[i].config.hostnames.geoKlb
+				if _, ok := geoOwnerMatcher[geoCode]; !ok {
+					geoOwnerMatcher[geoCode] = []types.GomegaMatcher{
 						ContainSubstring("heritage=external-dns,external-dns/owner="),
-						ContainSubstring(dnsRecord1.Status.OwnerID),
-						ContainSubstring(dnsRecord2.Status.OwnerID),
-					)),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
+					}
+				}
+				geoOwnerMatcher[geoCode] = append(geoOwnerMatcher[geoCode], ContainSubstring(ownerID))
+			}
+
+			By("[Common] checking common endpoints")
+			// A CNAME record for testHostname should always exist and be owned by all endpoints
+			By("[Common] checking " + testHostname + " endpoint")
+			Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+				"DNSName":       Equal(testHostname),
+				"Targets":       ConsistOf(testRecords[0].config.hostnames.klb),
+				"RecordType":    Equal("CNAME"),
+				"SetIdentifier": Equal(""),
+				"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+			}))))
+			totalEndpointsChecked++
+			By("[Common] checking " + testHostname + " TXT owner endpoint")
+			Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+				"DNSName":       Equal("kuadrant-cname-" + testHostname),
+				"Targets":       ContainElement(And(allOwnerMatcher...)),
+				"RecordType":    Equal("TXT"),
+				"SetIdentifier": Equal(""),
+				"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+			}))))
+			totalEndpointsChecked++
+
+			By("[Geo] checking geo endpoints")
+			if testDNSProvider == "google" {
+				// A CNAME record for klbHostName should always exist, be owned by all endpoints and target all geo hostnames
+				klbHostName := testRecords[0].config.hostnames.klb
+
+				allKlbGeoHostnames := []string{}
+				gcpGeoProps := []externaldnsendpoint.ProviderSpecificProperty{
+					{Name: "routingpolicy", Value: "geo"},
+				}
+				for g, h := range geoKlbHostname {
+					allKlbGeoHostnames = append(allKlbGeoHostnames, h)
+					gcpGeoProps = append(gcpGeoProps, externaldnsendpoint.ProviderSpecificProperty{Name: h, Value: g})
+				}
+
+				By("[Geo] checking " + klbHostName + " endpoint")
 				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName": Equal("kuadrant-cname-" + klbHostName),
-					"Targets": ContainElement(And(
-						ContainSubstring("heritage=external-dns,external-dns/owner="),
-						ContainSubstring(dnsRecord1.Status.OwnerID),
-						ContainSubstring(dnsRecord2.Status.OwnerID),
-					)),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+					"DNSName":          Equal(klbHostName),
+					"Targets":          ConsistOf(allKlbGeoHostnames),
+					"RecordType":       Equal("CNAME"),
+					"SetIdentifier":    Equal(""),
+					"RecordTTL":        Equal(externaldnsendpoint.TTL(300)),
+					"ProviderSpecific": ContainElements(gcpGeoProps),
 				}))))
+				totalEndpointsChecked++
+				By("[Geo] checking " + klbHostName + " TXT owner endpoint")
 				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-cname-" + geo1KlbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord1.Status.OwnerID + "\""),
+					"DNSName":       Equal("kuadrant-cname-" + klbHostName),
+					"Targets":       ContainElement(And(allOwnerMatcher...)),
 					"RecordType":    Equal("TXT"),
 					"SetIdentifier": Equal(""),
 					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
 				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-cname-" + geo2KlbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord2.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-a-" + cluster1KlbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord1.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-a-" + cluster2KlbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord2.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
+				totalEndpointsChecked++
 			}
 			if testDNSProvider == "aws" {
-				Expect(zoneEndpoints).To(HaveLen(16))
-				//Main Records
-				By("checking endpoint " + testHostname)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(testHostname),
-					"Targets":       ConsistOf(klbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
-				By("checking endpoint " + klbHostName + " - " + geoCode1)
+				// A CNAME record for klbHostName should exist for each geo and be owned by all endpoints in that geo
+				klbHostName := testRecords[0].config.hostnames.klb
+				for geoCode, geoRecords := range testGeoRecords {
+					geoKlbHostName := geoRecords[0].config.hostnames.geoKlb
+
+					By("[Geo] checking " + klbHostName + " -> " + geoCode + " -> " + geoKlbHostName + " - endpoint")
+
+					awsGeoCodeKey := "aws/geolocation-country-code"
+					if !provider.IsISO3166Alpha2Code(geoCode) {
+						awsGeoCodeKey = "aws/geolocation-continent-code"
+					}
+
+					Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+						"DNSName":       Equal(klbHostName),
+						"Targets":       ConsistOf(geoKlbHostName),
+						"RecordType":    Equal("CNAME"),
+						"SetIdentifier": Equal(geoCode),
+						"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+						"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
+							{Name: "alias", Value: "false"},
+							{Name: awsGeoCodeKey, Value: geoCode},
+						}),
+					}))))
+					totalEndpointsChecked++
+					By("[Geo] checking " + klbHostName + " -> " + geoCode + " - TXT owner endpoint")
+					Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+						"DNSName":       Equal("kuadrant-cname-" + klbHostName),
+						"Targets":       ContainElement(And(geoOwnerMatcher[geoCode]...)),
+						"RecordType":    Equal("TXT"),
+						"SetIdentifier": Equal(geoCode),
+						"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+						"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
+							{Name: awsGeoCodeKey, Value: geoCode},
+						}),
+					}))))
+					totalEndpointsChecked++
+				}
+
+				defaultGeoKlbHostName := testRecords[0].config.hostnames.defaultGeoKlb
+				defaultGeoCode := testRecords[0].config.testDefaultGeoCode
+
+				By("[Geo] checking endpoint " + klbHostName + " -> default")
 				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
 					"DNSName":       Equal(klbHostName),
-					"Targets":       ConsistOf(geo1KlbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(geoCode1),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-					"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
-						{Name: "alias", Value: "false"},
-						{Name: "aws/geolocation-country-code", Value: "US"},
-					}),
-				}))))
-				By("checking endpoint " + klbHostName + " - " + geoCode2)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(klbHostName),
-					"Targets":       ConsistOf(geo2KlbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(geoCode2),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-					"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
-						{Name: "alias", Value: "false"},
-						{Name: "aws/geolocation-continent-code", Value: "EU"},
-					}),
-				}))))
-				By("checking endpoint " + klbHostName + " - " + geoCode1)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(klbHostName),
-					"Targets":       ConsistOf(geo1KlbHostName),
+					"Targets":       ConsistOf(defaultGeoKlbHostName),
 					"RecordType":    Equal("CNAME"),
 					"SetIdentifier": Equal("default"),
 					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
@@ -616,82 +537,11 @@ var _ = Describe("Multi Record Test", func() {
 						{Name: "aws/geolocation-country-code", Value: "*"},
 					}),
 				}))))
-				By("checking endpoint " + geo1KlbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(geo1KlbHostName),
-					"Targets":       ConsistOf(cluster1KlbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(cluster1KlbHostName),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-					"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
-						{Name: "alias", Value: "false"},
-						{Name: "aws/weight", Value: "200"},
-					}),
-				}))))
-				By("checking endpoint " + geo2KlbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(geo2KlbHostName),
-					"Targets":       ConsistOf(cluster2KlbHostName),
-					"RecordType":    Equal("CNAME"),
-					"SetIdentifier": Equal(cluster2KlbHostName),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-					"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
-						{Name: "alias", Value: "false"},
-						{Name: "aws/weight", Value: "200"},
-					}),
-				}))))
-				By("checking endpoint " + cluster1KlbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(cluster1KlbHostName),
-					"Targets":       ConsistOf(testTargetIP1),
-					"RecordType":    Equal("A"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-				}))))
-				By("checking endpoint " + cluster2KlbHostName)
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal(cluster2KlbHostName),
-					"Targets":       ConsistOf(testTargetIP2),
-					"RecordType":    Equal("A"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
-				}))))
-				//Txt Records
-				By("checking TXT owner endpoints")
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName": Equal("kuadrant-cname-" + testHostname),
-					"Targets": ContainElement(And(
-						ContainSubstring("heritage=external-dns,external-dns/owner="),
-						ContainSubstring(dnsRecord1.Status.OwnerID),
-						ContainSubstring(dnsRecord2.Status.OwnerID),
-					)),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
+				totalEndpointsChecked++
+				By("[Geo] checking " + klbHostName + " -> default - TXT owner endpoint")
 				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
 					"DNSName":       Equal("kuadrant-cname-" + klbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord1.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(geoCode1),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-					"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
-						{Name: "aws/geolocation-country-code", Value: "US"},
-					}),
-				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-cname-" + klbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord2.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(geoCode2),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-					"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
-						{Name: "aws/geolocation-continent-code", Value: "EU"},
-					}),
-				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-cname-" + klbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord1.Status.OwnerID + "\""),
+					"Targets":       ContainElement(And(geoOwnerMatcher[defaultGeoCode]...)),
 					"RecordType":    Equal("TXT"),
 					"SetIdentifier": Equal("default"),
 					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
@@ -699,88 +549,127 @@ var _ = Describe("Multi Record Test", func() {
 						{Name: "aws/geolocation-country-code", Value: "*"},
 					}),
 				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-cname-" + geo1KlbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord1.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(cluster1KlbHostName),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-					"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
-						{Name: "aws/weight", Value: "200"},
-					}),
-				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-cname-" + geo2KlbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord2.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(cluster2KlbHostName),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-					"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
-						{Name: "aws/weight", Value: "200"},
-					}),
-				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-a-" + cluster1KlbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord1.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
-				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
-					"DNSName":       Equal("kuadrant-a-" + cluster2KlbHostName),
-					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + dnsRecord2.Status.OwnerID + "\""),
-					"RecordType":    Equal("TXT"),
-					"SetIdentifier": Equal(""),
-					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
-				}))))
+				totalEndpointsChecked++
 			}
 
-			By("ensuring the authoritative nameserver resolves the hostname")
-			// speed up things by using the authoritative nameserver
-			authoritativeResolver := ResolverForDomainName(testZoneDomainName)
+			By("[Weight] checking weighted endpoints")
+			if testDNSProvider == "google" {
+				// A weighted CNAME record should exist for each geo, be owned by all endpoints in that geo, and target the hostname of all clusters in that geo
+				for geoCode, geoRecords := range testGeoRecords {
+					geoKlbHostName := geoRecords[0].config.hostnames.geoKlb
+
+					allGeoClusterHostnames := []string{}
+					gcpWeightProps := []externaldnsendpoint.ProviderSpecificProperty{
+						{Name: "routingpolicy", Value: "weighted"},
+					}
+					for i := range geoRecords {
+						geoClusterHostname := geoRecords[i].config.hostnames.clusterKlb
+						allGeoClusterHostnames = append(allGeoClusterHostnames, geoClusterHostname)
+						gcpWeightProps = append(gcpWeightProps, externaldnsendpoint.ProviderSpecificProperty{Name: geoClusterHostname, Value: "200"})
+					}
+
+					By("[Weight] checking " + geoKlbHostName + " endpoint")
+					Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+						"DNSName":          Equal(geoKlbHostName),
+						"Targets":          ConsistOf(allGeoClusterHostnames),
+						"RecordType":       Equal("CNAME"),
+						"SetIdentifier":    Equal(""),
+						"RecordTTL":        Equal(externaldnsendpoint.TTL(60)),
+						"ProviderSpecific": ContainElements(gcpWeightProps),
+					}))))
+					totalEndpointsChecked++
+					By("[Weight] checking " + geoKlbHostName + " TXT owner endpoint")
+					Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+						"DNSName":       Equal("kuadrant-cname-" + geoKlbHostName),
+						"Targets":       ContainElement(And(geoOwnerMatcher[geoCode]...)),
+						"RecordType":    Equal("TXT"),
+						"SetIdentifier": Equal(""),
+						"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+					}))))
+					totalEndpointsChecked++
+				}
+			}
+			if testDNSProvider == "aws" {
+				// A weighted CNAME record should exist for each dns record in each geo and be owned only by that endpoint
+				for _, geoRecords := range testGeoRecords {
+					geoKlbHostName := geoRecords[0].config.hostnames.geoKlb
+					for i := range geoRecords {
+						clusterKlbHostName := geoRecords[i].config.hostnames.clusterKlb
+						ownerID := geoRecords[i].record.Status.OwnerID
+						By("[Weight] checking " + geoKlbHostName + " -> " + clusterKlbHostName + " - endpoint")
+						Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+							"DNSName":       Equal(geoKlbHostName),
+							"Targets":       ConsistOf(clusterKlbHostName),
+							"RecordType":    Equal("CNAME"),
+							"SetIdentifier": Equal(clusterKlbHostName),
+							"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
+							"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
+								{Name: "alias", Value: "false"},
+								{Name: "aws/weight", Value: "200"},
+							}),
+						}))))
+						totalEndpointsChecked++
+						By("[Weight] checking " + geoKlbHostName + " -> " + clusterKlbHostName + " -> " + ownerID + " TXT owner endpoint")
+						Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+							"DNSName":       Equal("kuadrant-cname-" + geoKlbHostName),
+							"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + ownerID + "\""),
+							"RecordType":    Equal("TXT"),
+							"SetIdentifier": Equal(clusterKlbHostName),
+							"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+							"ProviderSpecific": Equal(externaldnsendpoint.ProviderSpecific{
+								{Name: "aws/weight", Value: "200"},
+							}),
+						}))))
+						totalEndpointsChecked++
+					}
+				}
+			}
+
+			By("[Cluster] checking cluster endpoints")
+			// An A record with the cluster target IP should exist for each dns record and owned only by that endpoint
+			for i := range testRecords {
+				clusterKlbHostName := testRecords[i].config.hostnames.clusterKlb
+				clusterTargetIP := testRecords[i].config.testTargetIP
+				ownerID := testRecords[i].record.Status.OwnerID
+				By("[Cluster] checking " + clusterKlbHostName + " endpoint")
+				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+					"DNSName":       Equal(clusterKlbHostName),
+					"Targets":       ConsistOf(clusterTargetIP),
+					"RecordType":    Equal("A"),
+					"SetIdentifier": Equal(""),
+					"RecordTTL":     Equal(externaldnsendpoint.TTL(60)),
+				}))))
+				totalEndpointsChecked++
+				By("[Cluster] checking " + clusterKlbHostName + " TXT owner endpoint")
+				Expect(zoneEndpoints).To(ContainElement(PointTo(MatchFields(IgnoreExtras, Fields{
+					"DNSName":       Equal("kuadrant-a-" + clusterKlbHostName),
+					"Targets":       ConsistOf("\"heritage=external-dns,external-dns/owner=" + ownerID + "\""),
+					"RecordType":    Equal("TXT"),
+					"SetIdentifier": Equal(""),
+					"RecordTTL":     Equal(externaldnsendpoint.TTL(300)),
+				}))))
+				totalEndpointsChecked++
+			}
+
+			By("checking all endpoints were validated")
+			Expect(totalEndpointsChecked).To(Equal(expectedEndpointsLen))
+
+			By("deleting all remaining dns records")
+			for _, tr := range testRecords {
+				err := tr.cluster.k8sClient.Delete(ctx, tr.record,
+					client.PropagationPolicy(metav1.DeletePropagationForeground))
+				Expect(client.IgnoreNotFound(err)).ToNot(HaveOccurred())
+			}
+
+			By(fmt.Sprintf("checking all dns records are removed within %s", recordsRemovedMaxDuration))
 			Eventually(func(g Gomega, ctx context.Context) {
-				ips, err := authoritativeResolver.LookupHost(ctx, testHostname)
-				g.Expect(err).NotTo(HaveOccurred())
-				GinkgoWriter.Printf("[debug] ips: %v\n", ips)
-				g.Expect(ips).To(Or(ContainElement(testTargetIP1), ContainElement(testTargetIP2)))
-			}, 300*time.Second, 10*time.Second, ctx).Should(Succeed())
+				for _, tr := range testRecords {
+					err := tr.cluster.k8sClient.Get(ctx, client.ObjectKeyFromObject(tr.record), tr.record)
+					g.Expect(err).To(MatchError(ContainSubstring("not found")))
+				}
+			}, recordsRemovedMaxDuration, 5*time.Second, ctx).Should(Succeed())
 
-			By("deleting dnsrecord " + dnsRecord2.Name)
-			err = k8sClient.Delete(ctx, dnsRecord2,
-				client.PropagationPolicy(metav1.DeletePropagationForeground))
-			Expect(client.IgnoreNotFound(err)).ToNot(HaveOccurred())
-
-			By("checking dnsrecord " + dnsRecord2.Name + " is removed")
-			Eventually(func(g Gomega, ctx context.Context) {
-				err := k8sClient.Get(ctx, client.ObjectKeyFromObject(dnsRecord2), dnsRecord2)
-				g.Expect(err).To(HaveOccurred())
-				g.Expect(err).To(MatchError(ContainSubstring("not found")))
-			}, 10*time.Second, 1*time.Second, ctx).Should(Succeed())
-
-			By("ensuring the authoritative nameserver resolves the hostname")
-			Eventually(func(g Gomega, ctx context.Context) {
-				ips, err := authoritativeResolver.LookupHost(ctx, testHostname)
-				g.Expect(err).NotTo(HaveOccurred())
-				GinkgoWriter.Printf("[debug] ips: %v\n", ips)
-				g.Expect(ips).To(ConsistOf(testTargetIP1))
-			}, 300*time.Second, 10*time.Second, ctx).Should(Succeed())
-
-			By("ensuring zone records are updated as expected")
-			//ToDo mnairn Add more checks in here
-
-			By("deleting dnsrecord " + dnsRecord1.Name)
-			err = k8sClient.Delete(ctx, dnsRecord1,
-				client.PropagationPolicy(metav1.DeletePropagationForeground))
-			Expect(client.IgnoreNotFound(err)).ToNot(HaveOccurred())
-
-			By("checking dnsrecord " + dnsRecord1.Name + " is removed")
-			Eventually(func(g Gomega, ctx context.Context) {
-				err := k8sClient.Get(ctx, client.ObjectKeyFromObject(dnsRecord1), dnsRecord1)
-				g.Expect(err).To(HaveOccurred())
-				g.Expect(err).To(MatchError(ContainSubstring("not found")))
-			}, 10*time.Second, 1*time.Second, ctx).Should(Succeed())
-
-			By("ensuring zone records are all removed as expected")
+			By("checking provider zone records are all removed")
 			Eventually(func(g Gomega, ctx context.Context) {
 				zoneEndpoints, err := EndpointsForHost(ctx, testProvider, testHostname)
 				g.Expect(err).NotTo(HaveOccurred())
